+++ OpenGL-GameEngine/include/PauseState.h
@@ -13,7 +13,7 @@ public:
     ~PauseState() override;
 
     void Initialize(Application* app) override;
-    void Update(Application* app, double dt) override;
+    void Update(Application* app, float dt) override;
     void Render(Application* app) override;
     bool isOpaque() const override { return false; }
