+++ OpenGL-GameEngine/include/PlayState.h
@@ -21,7 +21,7 @@ public:
 
     // State Interface Overrides
     void Initialize(Application* app) override;
-    void Update(Application* app, double dt) override;
+    void Update(Application* app, float dt) override;
     void Render(Application* app) override;
     bool isOpaque() const override { return true; }
