+++ OpenGL-GameEngine/include/State.h
@@ -10,7 +10,7 @@ public:
     virtual void Initialize(Application* app) = 0;
 
     // Called every frame (logic & input)
-    virtual void Update(Application* app, double dt) = 0;
+    virtual void Update(Application* app, float dt) = 0;
 
     // Called every frame (graphics)
     virtual void Render(Application* app) = 0;
